Refactor redux
